requirements.txt
ultralytics==8.0.227
opencv-python-headless==4.8.1.78
pandas==2.1.1
matplotlib==3.8.0
seaborn==0.13.0
streamlit==1.26.0
schedule==1.2.0
scikit-learn==1.3.0
statsmodels==0.14.0
prophet==1.1.5
google-api-python-client==2.101.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
ipywidgets